Fix mobile app theme bug: updated  - Updated backend logic in gemini.py
